+++ ravif/src/lib.rs
@@ -121,7 +121,7 @@ pub type Result<T> = core::result::Result<T, At<Error>>;
 #[deprecated = "Renamed to `ColorModel`"]
 pub type ColorSpace = ColorModel;
 
-pub use animated::{AnimFrame, AnimFrame16, AnimFrameRgba, AnimFrameRgba16, EncodedAnimation};
+pub use animated::{AnimFrame, AnimFrame16, AnimFrameRgba, AnimFrameRgba16, TimedAnimFrame, EncodedAnimation};
 pub use av1encoder::{AlphaColorMode, BitDepth, ChromaSubsampling, EncodedImage, Encoder, GainMapData, DEFAULT_MAX_PIXELS};
 #[cfg(feature = "imazen")]
 pub use av1encoder::FRAME_HINTS_LIVE;